Adding full contact section
Here I added a contact me section with one link to my github profile using an href, as well as acontact details for my phone number and email. I think this was the correct way of doing it, however the target section in the github section I wasn't positive of. I added a class to each of these and put them inside the body/html tags. Correct me if I'm wrong!