Added link to first project
Here I added a link to my project I created in Codepen.io. I tagged it with a class of projects grid so I can add more in the future and it will look nice.  I also am looking to add an image for the survey form but am unsure this is the one I want to stick to.